# api/openapi.yaml
openapi: 3.0.3
info:
  title: RSP API
  version: 2.0.0
  description: Same-origin API for the Ravi's Study Program application.
servers:
  - url: /api/v2
security:
  - bearerAuth: []
paths:
...
  /openapi.json:
    get:
      operationId: getOpenAPI
      security: []
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          description: This OpenAPI document as JSON
          content:
            {
              application/json:
                { schema: { type: object, additionalProperties: true } },
            }
  /me:
    get:
      operationId: getMe
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Current user and effective programme access,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/Me' } },
              },
          }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
    patch:
      operationId: updateMe
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: '#/components/schemas/UpdateMe' } },
            },
        }
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Updated current user,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/Me' } },
              },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /me/slug-suggestion:
    get:
      operationId: suggestMeSlug
      description: Returns a server-generated profile slug candidate that is unique at response time. PATCH /me remains authoritative and returns 409 if a concurrent claim wins.
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Unique profile slug candidate,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/SlugSuggestion' } },
              },
          }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /me/practice-settings:
    get:
      operationId: getPracticeSettings
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Current member practice preferences,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/PracticeSettings' } },
              },
          }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
    patch:
      operationId: updatePracticeSettings
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: '#/components/schemas/PracticeSettingsMutation' },
                },
            },
        }
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Updated practice preferences,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/PracticeSettings' } },
              },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /users:
    get:
      operationId: listUsers
      parameters:
        [
          { $ref: '#/components/parameters/Limit' },
          { $ref: '#/components/parameters/Cursor' },
          { $ref: '#/components/parameters/Direction' },
          { $ref: '#/components/parameters/Sort' },
          { name: query, in: query, schema: { type: string, maxLength: 100 } },
          {
            name: seasonRole,
            in: query,
            schema: { $ref: '#/components/schemas/SeasonRole' },
          },
          {
            name: globalRole,
            in: query,
            schema: { $ref: '#/components/schemas/GlobalRole' },
          },
        ]
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: User page,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/UserPage' } },
              },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /users/{id}:
    parameters:
      [
        {
          name: id,
          in: path,
          required: true,
          description: Opaque user ID or unique profile slug,
          schema: { type: string },
        },
      ]
    get:
      operationId: getUser
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Member profile. Private fields are included only for the member or an authorised relationship.,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        oneOf:
                          [
                            { $ref: '#/components/schemas/User' },
                            { $ref: '#/components/schemas/UserPrivate' },
                          ],
                      },
                  },
              },
          }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /users/{id}/practice-goals/enable:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    post:
      operationId: enableUserPracticeGoals
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: '#/components/schemas/PracticeGoalsEnablement' },
                },
            },
        }
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Personal practice goals enabled,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/PracticeSettings' } },
              },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /users/{id}/practice-settings:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    get:
      operationId: getUserPracticeSettings
      description: Returns private practice settings, including the independent settings revision, to the member, their assigned mentor, their season Coordinator, a Director, or a System Admin.
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Member practice settings,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/PracticeSettings' } },
              },
          }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /seasons:
    get:
      operationId: listSeasons
      parameters:
        [
          { $ref: '#/components/parameters/Limit' },
          { $ref: '#/components/parameters/Cursor' },
          { $ref: '#/components/parameters/Direction' },
          { $ref: '#/components/parameters/Sort' },
          {
            name: status,
            in: query,
            schema: { type: string, enum: [open, closed] },
          },
        ]
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Season page,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/SeasonPage' } },
              },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      operationId: createSeason
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: '#/components/schemas/SeasonMutation' } },
            },
        }
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '201':
          {
            description: Created season,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/Season' } },
              },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /seasons/{id}:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    get:
      operationId: getSeason
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Season,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/Season' } },
              },
          }
        '404': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
    patch:
      operationId: updateSeason
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: '#/components/schemas/SeasonUpdate' } },
            },
        }
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Updated season,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/Season' } },
              },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /seasons/{id}/close:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    post:
      operationId: closeSeason
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: '#/components/schemas/ReasonedRevision' } },
            },
        }
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Closed season,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/Season' } },
              },
          }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /seasons/{id}/resources:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    patch:
      operationId: updateSeasonResources
      description: Updates only the resource link for an open season. The season Coordinator, a Director, or a System Admin may call this operation with recent MFA.
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: '#/components/schemas/SeasonResourceUpdate' },
                },
            },
        }
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Updated season,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/Season' } },
              },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /seasons/{id}/enrollment-candidates:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    get:
      operationId: listEnrollmentCandidates
      description: Lists verified active accounts that are not yet affiliated with this season. Only season administrators with recent MFA may call this operation; no private account fields are returned.
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Direction' }
        - { $ref: '#/components/parameters/Sort' }
        - { name: query, in: query, schema: { type: string, maxLength: 100 } }
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Enrollment candidate page,
            content:
              {
                application/json:
                  {
                    schema:
                      { $ref: '#/components/schemas/EnrollmentCandidatePage' },
                  },
              },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /seasons/{id}/reopen:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    post:
      operationId: reopenSeason
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: '#/components/schemas/ReasonedRevision' } },
            },
        }
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Reopened season,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/Season' } },
              },
          }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /seasons/{id}/weeks:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    get:
      operationId: listSeasonWeeks
      parameters:
        [
          { $ref: '#/components/parameters/Limit' },
          { $ref: '#/components/parameters/Cursor' },
          { $ref: '#/components/parameters/Direction' },
          { $ref: '#/components/parameters/Sort' },
        ]
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Week page,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/WeekPage' } },
              },
          }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      operationId: createSeasonWeek
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: '#/components/schemas/WeekMutation' } },
            },
        }
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '201':
          {
            description: Created week,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/Week' } },
              },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /recommendations/current:
    get:
      operationId: getCurrentRecommendation
      responses:
        default: { $ref: '#/components/responses/Problem' }
        '200':
          {
            description: Active recommendation,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/Recommendation' } },
              },
          }
        '404': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    Cursor: { name: cursor, in: query, schema: { type: string } }
...
  responses:
...
    Problem:
      description: Request failed
      content:
        {
          application/problem+json:
            { schema: { $ref: '#/components/schemas/ProblemDetails' } },
        }
    RateLimited:
      description: Rate limited
      headers:
        {
          Retry-After:
            { required: true, schema: { type: integer, minimum: 1 } },
        }
      content:
        {
          application/problem+json:
            { schema: { $ref: '#/components/schemas/ProblemDetails' } },
        }
  schemas:
...
    Me:
      allOf:
        - $ref: '#/components/schemas/UserPrivate'
        - type: object
          required: [emailVerified, mfaVerified, seasonRoles, alumni]
          properties:
            emailVerified: { type: boolean }
            mfaVerified: { type: boolean }
            seasonRoles:
              {
                type: array,
                items: { $ref: '#/components/schemas/MeSeasonRole' },
              }
            alumni: { type: boolean }
    UpdateMe:
      type: object
      additionalProperties: false
      required: [name, timezone, revision]
      properties:
        {
          name: { type: string, minLength: 1, maxLength: 100 },
          slug:
            { type: string, pattern: '^[a-z0-9](?:[a-z0-9-]{1,48}[a-z0-9])?$' },
          avatarUrl: { type: string, format: uri, nullable: true },
          timezone: { type: string },
          revision: { $ref: '#/components/schemas/Revision' },
        }
...